@@ -13,31 +13,35 @@ class _IndexCurpertinoBottomNavBarState extends State<IndexCurpertinoBottomNavBa
   Widget build(BuildContext context) {
     return Scaffold(
       body: IndexedStack(
-        children: <Widget>[
-          PageOne(),
-          PageTwo(),
-          PageThree()
-        ],
+        children: <Widget>[PageOne(), PageTwo(), PageThree()],
         index: page,
       ),
       bottomNavigationBar: CupertinoTabBar(
           items: [
             BottomNavigationBarItem(
-                icon: Icon(Icons.home, color: page == 0 ? Colors.blue : Colors.grey,),
-                title: Text('基础', style: TextStyle(color: page == 0 ? Colors.blue : Colors.grey,),)),
+                icon: Icon(
+                  Icons.home,
+                  color: page == 0 ? Colors.blue : Colors.grey,
+                ),
+                label: '基础'),
             BottomNavigationBarItem(
-                icon: Icon(Icons.message, color: page == 1 ? Colors.blue : Colors.grey,),
-                title: Text('插件', style: TextStyle(color: page == 1 ? Colors.blue : Colors.grey,),)),
+                icon: Icon(
+                  Icons.message,
+                  color: page == 1 ? Colors.blue : Colors.grey,
+                ),
+                label: '插件'),
             BottomNavigationBarItem(
-                icon: Icon(Icons.apps, color: page == 2 ? Colors.blue : Colors.grey,),
-                title: Text('iOS风格', style: TextStyle(color: page == 3 ? Colors.blue : Colors.grey,),)),
+                icon: Icon(
+                  Icons.apps,
+                  color: page == 2 ? Colors.blue : Colors.grey,
+                ),
+                label: 'iOS风格'),
           ],
           onTap: (index) {
             setState(() {
               page = index;
             });
-          }
-      ),
+          }),
     );
   }
 }
