@@ -1,6 +1,6 @@
 //import 'package:flutter/material.dart';
 //
-//import 'package:video_player/video_player.dart';
+import 'package:video_player/video_player.dart';
 //
 //class NetworkPlayerLifeCycle extends PlayerLifeCycle {
 //  NetworkPlayerLifeCycle(String dataSource, VideoWidgetBuilder childBuilder)
